División de Cotización dolar en fragment y viewModel (#1)
* Notification Added (#16)

* Update HomeFragment.kt

* Separate fragment and viewmodel

---------
